feat: render front blocks
